fixup models delete cruft directory
