+++ src/edu/wgu/scheduler/controllers/AppointmentViewController.java
@@ -16,7 +16,6 @@ import javafx.scene.layout.VBox;
 import javafx.scene.layout.HBox;
 import sun.reflect.generics.reflectiveObjects.NotImplementedException;
 
-import javax.xml.soap.Text;
 import java.io.IOException;
 import java.net.URL;
 import java.sql.*;
@@ -260,21 +259,21 @@ public class AppointmentViewController implements Initializable {
 
     private void setupCollections() {
         appointments = FXCollections.observableList(new LinkedList<>(), (AppointmentProperty ap) -> new Observable[]{
-                ap.title(),
-                ap.description(),
-                ap.location(),
-                ap.contact(),
-                ap.url(),
-                ap.start(),
-                ap.end(),
-                ap.customerId()
+                ap.titleProperty(),
+                ap.descriptionProperty(),
+                ap.locationProperty(),
+                ap.contactProperty(),
+                ap.urlProperty(),
+                ap.startProperty(),
+                ap.endProperty(),
+                ap.customerIdProperty()
         });
 
         reminders = FXCollections.observableList(new LinkedList<>(), re -> new Observable[]{
-                re.remindercol(),
+                re.remindercolProperty(),
                 re.reminderDate(),
-                re.snoozeIncrement(),
-                re.snoozeIncrementTypeId()
+                re.snoozeIncrementProperty(),
+                re.snoozeIncrementTypeIdProperty()
         });
     }
 
@@ -307,31 +306,31 @@ public class AppointmentViewController implements Initializable {
             PreparedStatement statement = connection.prepareStatement("SELECT * FROM Appointments");
             ResultSet rs = statement.executeQuery();
             while (rs.next()){
-                edu.wgu.scheduler.models.AppointmentView app = new AppointmentView(rs.getString("title"),
-                        rs.getString("description"),
-                        rs.getString("location"),
-                        rs.getString("contact"),
-                        rs.getString("url"),
-                        rs.getString("customerName"),
-                        rs.getTimestamp("start"),
-                        rs.getTimestamp("end"),
-                        rs.getDate("createDate"),
-                        rs.getString("createdBy"),
-                        rs.getTimestamp("lastUpdate"));
+                AppointmentViewProperty.AppointmentView app = new AppointmentViewProperty.AppointmentView(rs.getString("titleProperty"),
+                                                                                                          rs.getString("descriptionProperty"),
+                                                                                                          rs.getString("locationProperty"),
+                                                                                                          rs.getString("contactProperty"),
+                                                                                                          rs.getString("urlProperty"),
+                                                                                                          rs.getString("customerName"),
+                                                                                                          rs.getTimestamp("startProperty"),
+                                                                                                          rs.getTimestamp("endProperty"),
+                                                                                                          rs.getDate("createDate"),
+                                                                                                          rs.getString("createdBy"),
+                                                                                                          rs.getTimestamp("lastUpdate"));
                 appointmentViews.add(new AppointmentViewProperty(app));
             }
 
             statement = connection.prepareStatement("SELECT * FROM customer");
             rs = statement.executeQuery();
             while(rs.next()){
-                CustomerProperty.Customer cu = new CustomerProperty.Customer(
+                Customer cu = new Customer(
                         rs.getDate("createDate").toLocalDate(),
-                        rs.getInt("customerId"),
+                        rs.getInt("customerIdProperty"),
                         rs.getByte("active"),
                         rs.getInt("addressId"),
                         rs.getString("createdBy"),
                         rs.getString("customerName"),
-                        rs.getString("lastUpdatedBy"),
+                        rs.getString("lastUpdatedByProperty"),
                         rs.getTimestamp("lastUpdate")
                 );
                 customers.add(new CustomerProperty(cu));
@@ -343,17 +342,17 @@ public class AppointmentViewController implements Initializable {
                 Appointment ap = new Appointment(
                         rs.getDate("createDate").toLocalDate(),
                         rs.getInt("appointmentId"),
-                        rs.getString("contact"),
+                        rs.getString("contactProperty"),
                         rs.getString("createdBy"),
-                        rs.getInt("customerId"),
-                        rs.getString("description"),
-                        rs.getString("end"),
+                        rs.getInt("customerIdProperty"),
+                        rs.getString("descriptionProperty"),
+                        rs.getString("endProperty"),
                         rs.getTimestamp("lastUpdate"),
-                        rs.getString("lastUpdatedBy"),
-                        rs.getString("location"),
-                        rs.getString("start"),
-                        rs.getString("title"),
-                        rs.getString("url")
+                        rs.getString("lastUpdatedByProperty"),
+                        rs.getString("locationProperty"),
+                        rs.getString("startProperty"),
+                        rs.getString("titleProperty"),
+                        rs.getString("urlProperty")
                 );
                 appointments.add(new AppointmentProperty(ap));
             }
@@ -367,9 +366,9 @@ public class AppointmentViewController implements Initializable {
                         rs.getString("createdBy"),
                         rs.getDate("createdDate").toLocalDate(),
                         ZonedDateTime.ofInstant(rs.getDate("reminderDate").toInstant(), ZoneId.systemDefault()),
-                        rs.getString("remindercol"),
-                        rs.getInt("snoozeIncrement"),
-                        rs.getInt("snoozeIncrementTypeId"));
+                        rs.getString("remindercolProperty"),
+                        rs.getInt("snoozeIncrementProperty"),
+                        rs.getInt("snoozeIncrementTypeIdProperty"));
                 reminders.add(new ReminderProperty(re));
             }
 
@@ -518,13 +517,13 @@ public class AppointmentViewController implements Initializable {
         }
 
         Appointment app = new Appointment(txtCreatedBy.getText(),
-                customerId,
-                txtDescription.getText(),
-                ZonedDateTime.of(LocalDateTime.parse(txtEnd.getText()), ZoneId.systemDefault()),
-                txtLocation.getText(),
-                ZonedDateTime.of(LocalDateTime.parse(txtStart.getText()), ZoneId.systemDefault()),
-                txtTitle.getText(),
-                txtUrl.getText());
+                                          customerId,
+                                          txtDescription.getText(),
+                                          ZonedDateTime.of(LocalDateTime.parse(txtEnd.getText()), ZoneId.systemDefault()),
+                                          txtLocation.getText(),
+                                          ZonedDateTime.of(LocalDateTime.parse(txtStart.getText()), ZoneId.systemDefault()),
+                                          txtTitle.getText(),
+                                          txtUrl.getText());
 
         try (Connection connection = dataSource.getConnection()) {
             PreparedStatement statement = connection.prepareStatement("UPDATE appointment\n" +
