+++ src/edu/wgu/scheduler/models/AppointmentViewProperty.java
@@ -1,8 +1,6 @@
 package edu.wgu.scheduler.models;
 
-import javafx.beans.property.ReadOnlyProperty;
-import javafx.beans.property.ReadOnlyStringProperty;
-import javafx.beans.property.SimpleObjectProperty;
+import javafx.beans.property.*;
 
 import java.sql.Date;
 import java.sql.Timestamp;
@@ -16,7 +14,7 @@ import java.time.ZonedDateTime;
  * <p>
  * Student ID: 000292065
  */
-public class AppointmentViewProperty extends SimpleObjectProperty<AppointmentView> implements IAppointmentView {
+public class AppointmentViewProperty extends SimpleObjectProperty<AppointmentViewProperty.AppointmentView> implements IAppointmentView {
 
     public AppointmentViewProperty(String title, String description, String location, String contact, String url, String customerName, Timestamp start, Timestamp end, Date createDate, String createdBy, Timestamp lastUpdate) {
         super(new AppointmentView(title, description, location, contact, url, customerName, start, end, createDate, createdBy, lastUpdate));
@@ -150,5 +148,139 @@ public class AppointmentViewProperty extends SimpleObjectProperty<AppointmentVie
     @Override
     public boolean equals(Object obj) {
         return getValue().equals(obj);
+    }
+
+    /**
+     * Created by Ian Cornett - icornet@wgu.edu on 7/30/2017 at 00:43.
+     * Part of Project: scheduler
+     * <p>
+     * Student ID: 000292065
+     */
+    public static class AppointmentView implements IAppointmentView {
+        private final Date createdDate;
+        private final String createdBy;
+        // Interface needs these components
+        private ReadOnlyStringProperty title;
+        private ReadOnlyStringProperty description;
+        private ReadOnlyStringProperty location;
+        private ReadOnlyStringProperty contact;
+        private ReadOnlyStringProperty url;
+        private ReadOnlyStringProperty customerName;
+        private ReadOnlyObjectProperty<Timestamp> start;
+        private ReadOnlyObjectProperty<Timestamp> end;
+        private ReadOnlyProperty<Timestamp> lastUpdated;
+
+        public AppointmentView(String title, String description, String location, String contact, String url, String customerName, Timestamp start, Timestamp end, Date createDate, String createdBy, Timestamp lastUpdate) {
+            this.title = new SimpleStringProperty(title);
+            this.description = new SimpleStringProperty(description);
+            this.location = new SimpleStringProperty(location);
+            this.contact = new SimpleStringProperty(contact);
+            this.customerName = new SimpleStringProperty(customerName);
+            this.start = new SimpleObjectProperty<>(start);
+            this.end = new SimpleObjectProperty<>(end);
+            this.lastUpdated = new SimpleObjectProperty<>(lastUpdate);
+            this.url = new SimpleStringProperty(url);
+            this.createdDate = createDate;
+            this.createdBy = createdBy;
+
+        }
+
+        public String getTitle() {
+            return title.getValue();
+        }
+
+        ReadOnlyStringProperty titleProperty() {
+            return title;
+        }
+
+        public String getDescription() {
+            return description.getValue();
+        }
+
+        ReadOnlyStringProperty descriptionProperty() {
+            return description;
+        }
+
+        public String getLocation() {
+            return location.getValue();
+        }
+
+        ReadOnlyStringProperty locationProperty() {
+            return location;
+        }
+
+        public String getContact() {
+            return contact.getValue();
+        }
+
+        ReadOnlyStringProperty contactProperty() {
+            return contact;
+        }
+
+        public String getUrl() {
+            return url.getValueSafe();
+        }
+
+        ReadOnlyStringProperty urlProperty() {
+            return url;
+        }
+
+        public String getCustomerName() {
+            return customerName.getValue();
+        }
+
+        ReadOnlyStringProperty customerNameProperty() {
+            return customerName;
+        }
+
+        public ZonedDateTime getStart() {
+            return ZonedDateTime.ofInstant(start.getValue().toInstant(), ZoneId.of("UTC"));
+        }
+
+        ReadOnlyProperty<Timestamp> startProperty() {
+            return start;
+        }
+
+        public ZonedDateTime getEnd() {
+            return ZonedDateTime.ofInstant(end.getValue().toInstant(), ZoneId.of("UTC"));
+        }
+
+        ReadOnlyProperty<Timestamp> endProperty() {
+            return end;
+        }
+
+        public LocalDate getCreateDate() {
+            return createdDate.toLocalDate();
+        }
+
+        public String getCreatedBy() {
+            return createdBy;
+        }
+
+        public ZonedDateTime getLastUpdate() {
+            return ZonedDateTime.ofInstant(lastUpdated.getValue().toInstant(), ZoneId.of("UTC"));
+        }
+
+        ReadOnlyProperty<Timestamp> lastUpdatedProperty() {
+            return lastUpdated;
+        }
+
+        @Override
+        public String toString() {
+            final StringBuffer sb = new StringBuffer("AppointmentView{");
+            sb.append("createdDate=").append(createdDate);
+            sb.append(", createdBy='").append(createdBy).append('\'');
+            sb.append(", titleProperty=").append(title);
+            sb.append(", descriptionProperty=").append(description);
+            sb.append(", locationProperty=").append(location);
+            sb.append(", contactProperty=").append(contact);
+            sb.append(", urlProperty=").append(url);
+            sb.append(", customerName=").append(customerName);
+            sb.append(", startProperty=").append(start);
+            sb.append(", endProperty=").append(end);
+            sb.append(", lastUpdated=").append(lastUpdated);
+            sb.append('}');
+            return sb.toString();
+        }
     }
 }
