+++ src/edu/wgu/scheduler/models/CityProperty.java
@@ -1,13 +1,8 @@
 package edu.wgu.scheduler.models;
 
-import com.sun.istack.internal.NotNull;
-import javafx.beans.property.ObjectProperty;
-import javafx.beans.property.SimpleIntegerProperty;
-import javafx.beans.property.SimpleObjectProperty;
-import javafx.beans.property.StringProperty;
+import javafx.beans.property.*;
 
 import java.sql.Timestamp;
-import java.time.ZoneId;
 import java.time.ZonedDateTime;
 
 /**
@@ -16,7 +11,7 @@ import java.time.ZonedDateTime;
  * <p>
  * Student ID: 000292065
  */
-public class CityProperty extends SimpleObjectProperty<CityProperty.City> implements ICity {
+public class CityProperty extends SimpleObjectProperty<City> implements ICity {
 
     public CityProperty(String city, int countryId) {
         super(new City(city, countryId));
@@ -131,139 +126,4 @@ public class CityProperty extends SimpleObjectProperty<CityProperty.City> implem
         init();
     }
 
-    /**
-     * Created by Ian Cornett - icornet@wgu.edu on 6/25/17.
-     * Part of Project: Java-8-SE-Programmer-II
-     * <p>
-     * Student ID: 000292065
-     */
-    public static class City implements ICity {
-        @NotNull
-        private final ZonedDateTime createDate;
-        private int cityId;
-        @NotNull
-        private StringProperty city;
-        @NotNull
-        private SimpleIntegerProperty countryId;
-        private String createdBy;
-        private ObjectProperty<ZonedDateTime> lastUpdate;
-        private StringProperty lastUpdateBy;
-
-        public City(String city, int countryId) {
-            this(city, countryId, null);
-        }
-
-        public City(String city, int countryId, String createdBy) {
-            this.city.setValue(city);
-            this.countryId.setValue(countryId);
-            this.createdBy = createdBy;
-            this.lastUpdateBy.setValue(createdBy);
-            this.createDate = ZonedDateTime.now(ZoneId.of("UTC"));
-        }
-
-        public City(int cityId, String city, int countryId, String createdBy, Timestamp lastUpdate, String lastUpdateBy, ZonedDateTime createDate) {
-            this.cityId = cityId;
-            this.city.setValue(city);
-            this.countryId.setValue(countryId);
-            this.createdBy = createdBy;
-            this.lastUpdate.setValue(ZonedDateTime.ofInstant(lastUpdate.toInstant(), ZoneId.systemDefault()));
-            this.lastUpdateBy.setValue(lastUpdateBy);
-            this.createDate = createDate;
-        }
-
-        @Override
-        public int getCityId() {
-            return cityId;
-        }
-
-        @Override
-        public String getCity() {
-            return city.getValue();
-        }
-
-        @Override
-        public void setCity(String city) {
-            this.city.setValue(city);
-        }
-
-        public StringProperty cityProperty() {
-            return city;
-        }
-
-        @Override
-        public int getCountryId() {
-            return countryId.getValue();
-        }
-
-        public SimpleIntegerProperty countryIdProperty() {
-            return countryId;
-        }
-
-        @Override
-        public ZonedDateTime getCreateDate() {
-            return createDate;
-        }
-
-        @Override
-        public String getCreatedBy() {
-            return createdBy;
-        }
-
-        @Override
-        public String getLastUpdateBy() {
-            return lastUpdateBy.getValueSafe();
-        }
-
-        @Override
-        public void setLastUpdateBy(String lastUpdateBy) {
-            this.lastUpdateBy.setValue(lastUpdateBy);
-        }
-
-        public StringProperty lastUpdateByProperty() {
-            return lastUpdateBy;
-        }
-
-        @Override
-        public ZonedDateTime getLastUpdate() {
-            return lastUpdate.get();
-        }
-
-        @Override
-        public void setLastUpdate(Timestamp lastUpdate) {
-            this.lastUpdate.set(ZonedDateTime.ofInstant(lastUpdate.toInstant(), ZoneId.systemDefault()));
-        }
-
-        public ObjectProperty<ZonedDateTime> lastUpdateProperty() {
-            return lastUpdate;
-        }
-
-        @Override
-        public String toString() {
-            final StringBuilder sb = new StringBuilder("City{");
-            sb.append("createDate=").append(createDate);
-            sb.append(", cityId=").append(cityId);
-            sb.append(", city=").append(city);
-            sb.append(", countryId=").append(countryId);
-            sb.append(", createdBy='").append(createdBy).append('\'');
-            sb.append(", lastUpdate=").append(lastUpdate);
-            sb.append(", lastUpdateBy=").append(lastUpdateBy);
-            sb.append('}');
-            return sb.toString();
-        }
-
-        @Override
-        public boolean equals(Object o) {
-            if (this == o) return true;
-            if (o == null || getClass() != o.getClass()) return false;
-
-            City city = (City) o;
-
-            return getCityId() == city.getCityId();
-        }
-
-        @Override
-        public int hashCode() {
-            return getCityId();
-        }
-    }
 }
