--- src/main/resources/Scheduler_en_US.properties
@@ -1,28 +1,0 @@
-Welcome=Welcome to C195 Scheduler!
-Username=Username:
-Password=Password:
-Login=Login
-Register=Register
-Enter\ Username=Please enter your username
-Enter\ Password=Please enter your password
-AppointmentTab=Appointments
-CustomerTab=Customers
-ListTab=View as list
-TableTab=View as table
-FileMenu=_File
-EditMenu=_Edit
-HelpMenu=Help
-CloseMenuItem=Close
-CopyMenuItem=Copy
-PasteMenuItem=Paste
-AboutMenuItem=About
-LogMenuItem=View Log
-LoginFailed=Login Failed!
-LoginFailedMessage=The username and password combination was invalid!
-userExists=User Exists!
-userExistsError=Please pick an unique username and try again!
-nullUsername=Username cannot be null or empty!
-invalidPassword=Password length must be at least three characters and the last character cannot be whitespace.
-LoginSuccessTitle=Success!
-LoginSuccessMessage=Login succeeded!
-LoginTitle=Please login to the scheduler
